@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-HelloWorld::Application.config.secret_token = 'd27c20935b01bec00359290297b80958414fb52ec56bde717e1c59f1cae9d267ebe0c85e54956e9d6fac2896de69cfa4e0f2b6be824cb2edd8d74a5f899f0c6a'
+Lpweb::Application.config.secret_token = 'e8485066e1dd6eec0bc56b36a105108f9a91ae5110cc5bac8f9c87770e5192964334818145f5c3dd90c6fa0dc155e736dea9982b823a7862a43f3d507107c6dc'
